Make stackmap accessible
